Moved complex JSON structure declarations to the internal `JsonLoader`
This type is relatively safe to keep inside this package (only), and
its only responsibility is really behaving as anti-corruption layer
with `composer.json`, `composer.lock` and `installed.json`, so
this is a good fit.
