+++ src/ComposerRequireChecker/FileLocator/LocateComposerPackageSourceFiles.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace ComposerRequireChecker\FileLocator;
 
 use ArrayIterator;
+use ComposerRequireChecker\JsonLoader;
 use Generator;
-use Psl\Type;
 
 use function array_map;
 use function array_merge;
@@ -17,35 +17,8 @@ use function ltrim;
 use function str_replace;
 
 /**
- * @psalm-type ComposerConfig = array{vendor-dir?: string}
- * @psalm-type ComposerAutoload = array{
- *                 exclude-from-classmap?: list<string>,
- *                 classmap?: list<string>,
- *                 files?: list<string>,
- *                 psr-0?: array<string, string|list<string>>,
- *                 psr-4?: array<string, string|list<string>>
- *             }
- * @psalm-type ComposerPackageData = array{
- *                 name: string,
- *                 require?: array<string, string>,
- *                 autoload?: ComposerAutoload,
- *                 config?: ComposerConfig
- *             }
- * @psalm-type InstalledComposerPackageData = array{
- *                  name: string,
- *                  require?: array<string, string>,
- *                  autoload?: ComposerAutoload
- *              }
- * @psalm-type ComposerData = array{
- *                 name?: string,
- *                 require?: array<string, string>,
- *                 autoload?: ComposerAutoload,
- *                 config?: ComposerConfig,
- *                 packages?: list<ComposerPackageData>
- *             }
- * @psalm-type InstalledComposerData = array{
- *                 packages?: list<InstalledComposerPackageData>
- *             }
+ * @psalm-import-type ComposerAutoload from JsonLoader
+ * @psalm-import-type ComposerData from JsonLoader
  */
 final class LocateComposerPackageSourceFiles
 {
@@ -176,70 +149,5 @@ final class LocateComposerPackageSourceFiles
     private function extractFilesFromDirectory(string $directory, array|null $blacklist): Generator
     {
         yield from (new LocateAllFilesByExtension())->__invoke(new ArrayIterator([$directory]), '.php', $blacklist);
-    }
-
-    /** @return Type\TypeInterface<InstalledComposerData> */
-    public static function installedDataType(): Type\TypeInterface
-    {
-        $autoload = Type\shape([
-            'exclude-from-classmap' => Type\optional(Type\vec(Type\string())),
-            'classmap' => Type\optional(Type\vec(Type\string())),
-            'files' => Type\optional(Type\vec(Type\string())),
-            'psr-0' => Type\optional(Type\dict(
-                Type\string(),
-                Type\union(Type\string(), Type\vec(Type\string())),
-            )),
-            'psr-4' => Type\optional(Type\dict(
-                Type\string(),
-                Type\union(Type\string(), Type\vec(Type\string())),
-            )),
-        ], true);
-
-        $package = Type\shape([
-            'name' => Type\string(),
-            'require' => Type\optional(Type\dict(Type\string(), Type\string())),
-            'autoload' => Type\optional($autoload),
-        ]);
-
-        return Type\shape([
-            'packages' => Type\optional(Type\vec($package)),
-        ], true);
-    }
-
-    /** @return Type\TypeInterface<ComposerData> */
-    public static function composerDataType(): Type\TypeInterface
-    {
-        $composerConfig = Type\shape([
-            'vendor-dir' => Type\optional(Type\string()),
-        ], true);
-
-        $autoload = Type\shape([
-            'exclude-from-classmap' => Type\optional(Type\vec(Type\string())),
-            'classmap' => Type\optional(Type\vec(Type\string())),
-            'files' => Type\optional(Type\vec(Type\string())),
-            'psr-0' => Type\optional(Type\dict(
-                Type\string(),
-                Type\union(Type\string(), Type\vec(Type\string())),
-            )),
-            'psr-4' => Type\optional(Type\dict(
-                Type\string(),
-                Type\union(Type\string(), Type\vec(Type\string())),
-            )),
-        ], true);
-
-        $package = Type\shape([
-            'name' => Type\string(),
-            'require' => Type\optional(Type\dict(Type\string(), Type\string())),
-            'autoload' => Type\optional($autoload),
-            'config' => Type\optional($composerConfig),
-        ]);
-
-        return Type\shape([
-            'name' => Type\optional(Type\string()),
-            'require' => Type\optional(Type\dict(Type\string(), Type\string())),
-            'autoload' => Type\optional($autoload),
-            'config' => Type\optional($composerConfig),
-            'packages' => Type\optional(Type\vec($package)),
-        ], true);
     }
 }
